Move playlists to separate files
